Move `TopAppBarHost` to its own package
