--- financial-connections/src/main/java/com/stripe/android/financialconnections/navigation/TopAppBarHost.kt
@@ -1,3 +1,0 @@
-package com.stripe.android.financialconnections.navigation
-
-internal interface TopAppBarHost
