+++ financial-connections/src/main/java/com/stripe/android/financialconnections/navigation/topappbar/TopAppBarHost.kt
@@ -1,0 +1,3 @@
+package com.stripe.android.financialconnections.navigation.topappbar
+
+internal interface TopAppBarHost
